Extract raw ordered window index module
